bugfix: negative literals were generated as unsigned

+++ spec/nim/tests/texpr_io_eof.nim
@@ -10,6 +10,6 @@ test "ExprIoEof":
 
   check(r.substream1.one == uint32(1262698832))
   check(r.substream1.two == 0'u32)
-  check(r.substream2.one == uint32(4294914349'u32))
+  check(r.substream2.one == uint32(4294914349'u64))
   check(r.substream2.two == uint32(1262698832))
   discard
